// src/app/models/currently.ts
export interface Currently {
  apparentTemperature: number;
  cloudCover: number;
  deg: string;
  dewPoint: number;
  feelsLikeC: number;
  humidity: number;
  icon: string;
  nearestStormBearing: number;
  nearestStormDistance: number;
  ozone: number;
  precipIntensity: number;
  precipProbability: number;
  pressure: number;
  summary: string;
  tempC: number;
  temperature: number;
  time: number;
  uvIndex: number;
  visibility: number;
  windBearing: number;
  windGust: number;
  windKph: number;
  windSpeed: number;
}